added return point to sendAction, moved collapse handler to settings ui, added hideCategories function, changed if to switch statement
